+++ sh/dev.sh
@@ -10,8 +10,6 @@ npm run serve &
 if [ "$PROFILE" ]
 then
 echo '\npreparing automation...';
-tsc automation.ts --skipLibCheck;
 node automation.js;
-rm automation.js;
 echo '\nautomation finished.\n';
 fi