+++ docs/changelog.md
@@ -37,6 +37,7 @@ Ends October 14, 2020
 - update puppeteer screenshot to be 1920x1080.
 - add index.profiler.tsx to Jest's ignored files.
 - add ability to save puppeteer automation logs to a JSON file.
+- refactor puppeteer automation file to be a js file rather than a ts file.
 
 ## Sprint 4
 Ends September 16, 2020